use config to set log and monitored resource names
