+++ api/config/config_local.json
@@ -77,5 +77,9 @@
     "enableVpcFlowLogs": true,
     "enableVpcServicePerimeter": true,
     "enableNewAccountCreation": true
+  },
+  "actionAudit": {
+    "monitoredResourceName": "global",
+    "logName":"action-audit-test-3"
   }
 }
