--- api/src/main/java/org/pmiops/workbench/audit/ActionAuditConfig.java
@@ -1,15 +1,0 @@
-package org.pmiops.workbench.audit;
-
-import com.google.cloud.logging.Logging;
-import com.google.cloud.logging.LoggingOptions;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ActionAuditConfig {
-
-  @Bean
-  public Logging getCloudLogging() {
-    return LoggingOptions.getDefaultInstance().getService();
-  }
-}
