+++ api/src/main/java/org/pmiops/workbench/audit/ActionAuditEvent.java
@@ -1,187 +1,35 @@
 package org.pmiops.workbench.audit;
 
+import com.google.cloud.MonitoredResource;
+import com.google.cloud.logging.LogEntry;
+import com.google.cloud.logging.Payload.JsonPayload;
+import com.google.cloud.logging.Severity;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Optional;
 
-public class ActionAuditEvent extends AbstractAuditableEvent {
+public interface ActionAuditEvent {
 
-  private final long timestamp;
-  private final AgentType agentType;
-  private final long agentId;
-  private final String actionId;
-  private Optional<String> agentEmail;
-  private final ActionType actionType;
-  private final TargetType targetType;
-  private final Optional<String> targetPropertyMaybe;
-  private Optional<Long> targetIdMaybe;
-  private final Optional<String> previousValueMaybe;
-  private final Optional<String> newValueMaybe;
+  public abstract long timestamp();
 
-  public ActionAuditEvent(
-      long timestamp,
-      AgentType agentType,
-      long agentId,
-      Optional<String> agentEmail,
-      String actionId,
-      ActionType actionType,
-      TargetType targetType,
-      Optional<String> targetPropertyMaybe,
-      Optional<Long> targetIdMaybe,
-      Optional<String> previousValueMaybe,
-      Optional<String> newValueMaybe) {
-    this.timestamp = timestamp;
-    this.agentType = agentType;
-    this.agentId = agentId;
-    this.agentEmail = agentEmail;
-    this.actionId = actionId;
-    this.actionType = actionType;
-    this.targetType = targetType;
-    this.targetPropertyMaybe = targetPropertyMaybe;
-    this.targetIdMaybe = targetIdMaybe;
-    this.previousValueMaybe = previousValueMaybe;
-    this.newValueMaybe = newValueMaybe;
-  }
+  public abstract String actionId();
 
-  @Override
-  public long timestamp() {
-    return timestamp;
-  }
+  public abstract AgentType agentType();
 
-  @Override
-  public AgentType agentType() {
-    return agentType;
-  }
+  public abstract long agentId();
 
-  @Override
-  public long agentId() {
-    return agentId;
-  }
+  public abstract Optional<String> agentEmail();
 
-  @Override
-  public Optional<String> agentEmail() {
-    return agentEmail;
-  }
+  public abstract ActionType actionType();
 
-  @Override
-  public String actionId() {
-    return actionId;
-  }
+  public abstract TargetType targetType();
 
-  @Override
-  public ActionType actionType() {
-    return actionType;
-  }
+  public abstract Optional<Long> targetId();
 
-  @Override
-  public TargetType targetType() {
-    return targetType;
-  }
+  public abstract Optional<String> targetProperty();
 
-  @Override
-  public Optional<Long> targetId() {
-    return targetIdMaybe; // check if this is null somehow
-  }
+  public abstract Optional<String> previousValue();
 
-  @Override
-  public Optional<String> targetProperty() {
-    return targetPropertyMaybe;
-  }
+  public abstract Optional<String> newValue();
 
-  @Override
-  public Optional<String> previousValue() {
-    return previousValueMaybe;
-  }
-
-  @Override
-  public Optional<String> newValue() {
-    return newValueMaybe;
-  }
-
-  // This builder handles null inputs for optional fields, simplifying code at the call site.
-  // For example setAgentEmail() can be called safely with null (or just not called it's not known).
-  public static class Builder {
-
-    private long timestamp;
-    private AgentType agentType;
-    private long agentId;
-    private String agentEmail;
-    private String actionId;
-    private ActionType actionType;
-    private TargetType targetType;
-    private String targetProperty;
-    private Long targetId;
-    private String previousValue;
-    private String newValue;
-
-    public Builder() {}
-
-    public Builder setTimestamp(long timestamp) {
-      this.timestamp = timestamp;
-      return this;
-    }
-
-    public Builder setAgentType(AgentType agentType) {
-      this.agentType = agentType;
-      return this;
-    }
-
-    public Builder setAgentId(long agentId) {
-      this.agentId = agentId;
-      return this;
-    }
-
-    public Builder setAgentEmail(String agentEmail) {
-      this.agentEmail = agentEmail;
-      return this;
-    }
-
-    public Builder setActionId(String actionId) {
-      this.actionId = actionId;
-      return this;
-    }
-
-    public Builder setActionType(ActionType actionType) {
-      this.actionType = actionType;
-      return this;
-    }
-
-    public Builder setTargetType(TargetType targetType) {
-      this.targetType = targetType;
-      return this;
-    }
-
-    public Builder setTargetProperty(String targetProperty) {
-      this.targetProperty = targetProperty;
-      return this;
-    }
-
-    public Builder setTargetId(Long targetId) {
-      this.targetId = targetId;
-      return this;
-    }
-
-    public Builder setPreviousValue(String previousValue) {
-      this.previousValue = previousValue;
-      return this;
-    }
-
-    public Builder setNewValue(String newValue) {
-      this.newValue = newValue;
-      return this;
-    }
-
-    public ActionAuditEvent build() {
-      return new ActionAuditEvent(
-          timestamp,
-          agentType,
-          agentId,
-          Optional.ofNullable(agentEmail),
-          actionId,
-          actionType,
-          targetType,
-          Optional.ofNullable(targetProperty),
-          Optional.ofNullable(targetId),
-          Optional.ofNullable(previousValue),
-          Optional.ofNullable(newValue));
-    }
-  }
 }
