+++ api/src/test/java/org/pmiops/workbench/audit/ActionAuditServiceTest.java
@@ -1,8 +1,10 @@
 package org.pmiops.workbench.audit;
 
 import static com.google.common.truth.Truth.assertThat;
+import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.verify;
 
+import com.google.cloud.MonitoredResource;
 import com.google.cloud.logging.LogEntry;
 import com.google.cloud.logging.Logging;
 import com.google.cloud.logging.Payload;
@@ -12,12 +14,15 @@ import com.google.common.collect.ImmutableList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
+import javax.inject.Provider;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.Mock;
+import org.pmiops.workbench.config.WorkbenchConfig;
+import org.pmiops.workbench.config.WorkbenchConfig.ActionAuditConfig;
 import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
@@ -26,22 +31,33 @@ public class ActionAuditServiceTest {
   private static final long AGENT_ID_1 = 101L;
   private static final long AGENT_ID_2 = 102L;
   @Mock private Logging mockLogging;
+  @Mock private MonitoredResource mockMonitoredResource;
+  @Mock private Provider<WorkbenchConfig> mockConfigProvider;
   @Captor private ArgumentCaptor<List<LogEntry>> logEntryListCaptor;
 
-  private String actionId;
-  private ActionAuditEvent event1;
-  private ActionAuditEvent event2;
+  private ActionAuditEventImpl event1;
+  private ActionAuditEventImpl event2;
   private ActionAuditService actionAuditService;
 
   @Before
   public void setUp() {
-    actionAuditService = new ActionAuditServiceImpl(mockLogging);
-    actionId = ActionAuditService.newActionId();
+
+    final WorkbenchConfig.ActionAuditConfig actionAudit = new ActionAuditConfig();
+    actionAudit.logName = "log_path_1";
+    actionAudit.monitoredResourceName = "monitored_resource_name_1";
+    final WorkbenchConfig workbenchConfig = new WorkbenchConfig();
+    workbenchConfig.actionAudit = actionAudit;
+    doReturn(workbenchConfig).when(mockConfigProvider).get();
+
+    actionAuditService = new ActionAuditServiceImpl(mockLogging,
+        mockMonitoredResource,
+        mockConfigProvider);
+    final String actionId = ActionAuditService.newActionId();
 
     // ordinarily events sharing an action would have more things in common than this,
     // but the schema doesn't require it
     event1 =
-        new ActionAuditEvent.Builder()
+        new ActionAuditEventImpl.Builder()
             .setAgentEmail("a@b.co")
             .setTargetType(TargetType.DATASET)
             .setTargetId(1L)
@@ -55,7 +71,7 @@ public class ActionAuditServiceTest {
             .build();
 
     event2 =
-        new ActionAuditEvent.Builder()
+        new ActionAuditEventImpl.Builder()
             .setAgentEmail("f@b.co")
             .setTargetType(TargetType.DATASET)
             .setTargetId(2L)
@@ -73,11 +89,14 @@ public class ActionAuditServiceTest {
   public void testSendsSingleEvent() {
     actionAuditService.send(event1);
     verify(mockLogging).write(logEntryListCaptor.capture());
-    List<LogEntry> entryList = logEntryListCaptor.getValue();
+
+    final List<LogEntry> entryList = logEntryListCaptor.getValue();
     assertThat(entryList.size()).isEqualTo(1);
-    LogEntry entry = entryList.get(0);
+
+    final LogEntry entry = entryList.get(0);
     assertThat(entry.getPayload().getType()).isEqualTo(Payload.Type.JSON);
-    JsonPayload jsonPayload = entry.getPayload();
+
+    final JsonPayload jsonPayload = entry.getPayload();
     assertThat(jsonPayload.getDataAsMap().size()).isEqualTo(11);
 
     Map<String, Object> payloadMap = jsonPayload.getDataAsMap();
